rs/requirements.txt
flask
flask-cors
psycopg2-binary
surprise
numpy
pandas
scikit-learn